[update] rename package
